Form-data & strategy fixes
Frontend requested that data should be posted and put using form-data
instead of raw. That has been added for capability, environment and
status. I've added part of it to Strategy but that's not working as
Lenny is searching how to post a LocalDate.
A few small fixes for the strategy update.